we need to introduce some State to represent the contents of the input. State is data that is bound to the view - only when state changes will React rerender and update the view. State is the source of truth in your React apps.
We introduce state in functional components using the useState hook.  Hooks have to go at the beginning of a function and the useState hook returns a tuple containing the current value of the state and a function to update the state which we can destructure in one line.
When assigning attributes with expressions in React we always use attribute={expression}, not attribute="{expression}".
add an onChange handler (note the camel-case) to call the setLocationSearch function whenever the control changes. Again, we use {} to supply an arrow function expression.
So when the change handler fires the state will be set, which in turn triggers a rerender.
To map this into the view we need to have a tr row for every entry in the array, so we literally do a map:
Note how within the map arrow function we can return nested JSX, and that JSX can in turn embed a nested {} expression.
This interleaving of code and view is very useful. We still need to make one final change. With lists of elements React wants a unique key for each entry so that it can efficiently perform change tracking. Later we will use the location id, but for now we can simply use the index within the table: